good it of open market long

+++ src/instructions/getStopLossLongIx.ts
@@ -7,7 +7,8 @@ export async function getSetStopLossLongIx(
     input: SetStopLossLongInput
 ) {
 
-    input.stopLossLimitPrice = Math.round(Number(input.stopLossLimitPrice) * 10 ** PRICE_DECIMALS);
+    // Convert to big int
+    input.stopLossLimitPrice = BigInt(Math.round(Number(input.stopLossLimitPrice) * 10 ** PRICE_DECIMALS));
     const ix = getSetStopLossLongInstruction(input);
 
     return ix;
